Use actual netcup return values, add tests
Now Append, Set and Delete compare the actual
records returned by the netcup API to the
previous ones to figure out what changed
and return that.
Added some integration tests, that has to be set up
with env variables for the DNS API credentials.
Also adjusted field names to be conform with
naming convention for abbreviations (like DNS and API).